requirements.txt
Pillow
setproctitle

